Home 🏠

Functions card pre_title title page_header page_body value_box icon alert button page body topbar footer sidebar_menu sidebar_brand horizontal_menu menu_item menu_dropdown tab_items tab_item navbar_menu

News Changelog

menu_dropdown {tabler}	R Documentation

Create a Dropdown Menu Item

Description

Create a nav-item with a dropdown menu using columns like the example

Usage

menu_dropdown(text, icon = NULL, href = NULL, items = list())

Arguments

text 	 Title of the dropdown

icon 	 Icon name (optional)

href 	 Link for the dropdown toggle (optional)

items 	 A list of character vectors or tags representing dropdown links; each element can be a character vector of length 2: c(text, href)

Value

An HTML tag representing the dropdown nav item